move some functions into another file and rename auth to security
